Change wait_for_container into waiting for Django runserver to run
This change was made to ensure the interactive terminal stays open.
Previously the interactive terminal was shut down if an entrypoint
script was used with a Deno server startup etc. This ensures the normal
use of our intial commands have been ran first as Django server is ran
as the "command" for runserver container in docker-compose.yml.